requirements/dog.txt
numpy
python-can
